Extract pure protocol helpers into a protocol module
Audit cleanup item: ext/yrb_lite/src/lib.rs had grown past 900 lines, mixing the
FFI/Ruby glue with pure, no-Ruby logic. Split the highest-value boundary out:

- New ext/yrb_lite/src/protocol.rs holds the pure (no Ruby, no GVL, no unsafe)
  helpers -- classify_message, merged_doc_update, awareness_client_ids_in,
  update_is_ready, update_advances_doc, doc_has_pending -- plus the cargo unit
  tests, which only exercise these functions and is_safe_client_id.
- lib.rs now declares `mod protocol;` and imports them; it keeps the Ruby-facing
  surface (the magnus-wrapped RbDoc/RbAwareness, the nogvl/error/validation
  helpers, and init). 950 -> 720 lines.

Pure mechanical move: no behavior change. cargo fmt + clippy clean, cargo test
9/9, the full Ruby suite (doc/awareness/sync/robustness/interop/thread_safety)
all green after rake compile.

A further split of the Doc and Awareness impls into their own modules is left as
optional follow-up -- the pure-logic vs FFI boundary is the one that most
improves reviewability, and the impls are already clearly sectioned.
